Rename vpc_http to http
