+++ lib/ibm/cloud/sdk/vpc/http.rb
